+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- [yuyo.components.builder][] descriptor for adding a raw component builder to
-  a column without a callback.
+- [yuyo.components.builder][] (a class descriptor),
+  [ActionColumnExecutor.add_builder][yuyo.components.ActionColumnExecutor.add_builder], and
+  [ActionColumnExecutor.add_static_builder][yuyo.components.ActionColumnExecutor.add_static_builder]
+  functions and methods for adding a component builders to a column without a callback.
 
 ### Changed
 - [yuyo.pagination.Page][] is now exported top-level at `Yuyo.Page`.